+++ CHANGELOG.md
@@ -81,6 +81,7 @@ They may change or be removed in future versions.
 * Some commands cannot be run from 'Command List' (https://github.com/qupath/qupath/issues/1647)
 * ROIs with small pieces could look very 'pointy' when viewed at low resolution (https://github.com/qupath/qupath/pull/1681)
 * Create training image does not work properly for z-stacks / time series (https://github.com/qupath/qupath/issues/1701)
+* Channel colors are not stored properly when using saved display settings (https://github.com/qupath/qupath/issues/1726)
 
 ### API changes
 * New `Map<String, String> getMetadata()` method added to `PathObject`, `Project` and `ProjectImageEntry` (https://github.com/qupath/qupath/pull/1587)